Revert "restructure application"
