--- api/bankApp/__init__.py
@@ -1,4 +1,0 @@
-# from bankApp.setting import app
-# from bankApp.extention import db
-# from bankApp.models import app
-# from bankApp.routes import withdraw
