--- api/test_project.py
@@ -1,3 +1,0 @@
-from index import app, db
-import pytest
-
