+++ requirements.txt
@@ -14,4 +14,3 @@ SQLAlchemy==2.0.28
 typing_extensions==4.10.0
 Werkzeug==3.0.1
 python-dotenv==1.0.1
-pytest==8.1.1
